# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GR-plan API doc",
        "version": "0.1"
    },
    "paths": {
        "/api/produto": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Get list of produtos",
                "description": "Returns list of produtos",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Requets"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Store new produto",
                "description": "Returns produto data",
                "operationId": "produto.store",
                "requestBody": {
                    "description": "Oasse as informações do produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Freezer"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "texto longo"
                                    },
                                    "tensao": {
                                        "type": "string",
                                        "example": "220v"
                                    },
                                    "marca_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/produto/{id}": {
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Delete existing produto",
                "description": "Deletes a record and returns no content",
                "operationId": "produto.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Produto id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad requets"
                    }
                }
            }
        }
    }
}